requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.4.0
channels==4.1.0
channels-redis==4.2.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
Django==4.2.16
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-ckeditor==6.7.1
django-environ==0.12.0
django-filter==25.1
django-js-asset==3.1.2
django-model-utils==5.0.0
django-redis==6.0.0
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework-api-response==0.1.0
h11==0.16.0
kombu==5.5.4
Markdown==3.8.2
msgpack==1.1.1
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-crontab==3.3.0
python-dateutil==2.9.0.post0
redis==6.4.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
